requirements.txt
discord.py
python-dotenv
flask
lxml
beautifulsoup4